requirements.txt
django
django-smart-selects
django-multiselectfield
